requirements.txt
blessings==1.7
boto3==1.12.41
botocore==1.15.41
docutils==0.15.2
jmespath==0.9.5
pkg-resources==0.0.0
python-dateutil==2.8.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
